requirements.txt
streamlit>=0.68.0
numpy>=1.18.1
pillow>=8.0.0
matplotlib>=3.3.2
wordcloud>=1.8.1